GG/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
certifi==2018.1.18
chardet==3.0.4
diff-match-patch==20121119
Django==1.11.9
django-admin-bootstrapped==2.5.7
django-appconf==1.0.2
django-celery==3.2.2
django-crispy-forms==1.7.0
django-formtools==2.1
django-haystack==2.8.0
django-import-export==0.7.0
django-redis==4.8.0
django-reversion==2.0.13
docutils==0.14
et-xmlfile==1.0.1
future==0.16.0
google-apputils==0.4.2
gunicorn==19.7.1
httplib2==0.9.2
idna==2.6
jdcal==1.3
jieba==0.39
kombu==3.0.37
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.5.0
paho-mqtt==1.3.1
Pillow==5.0.0
pymongo==3.6.0
python-dateutil==2.7.0
python-gflags==3.1.2
python-memcached==1.58
pytz==2017.2
PyYAML==3.12
redis==2.10.6
requests==2.18.4
setuptools-scm==1.17.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
vine==1.1.4
Whoosh==2.7.4
xlrd==1.1.0
xlwt==1.3.0
